+++ spec/client_spec.rb
@@ -7,4 +7,11 @@ describe 'client' do
       redis.reconnect.should == redis
     end
   end
+
+  context '#connect' do
+    it 'connects' do
+      redis = MockRedis.new
+      redis.connect.should == redis
+    end
+  end
 end