+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # MockRedis Changelog
 
+### master (unreleased)
+
+* Fix bug where SETBIT command would not correctly unset a bit
+* Add support for the `connect` method
+
 ### 0.12.1
 
 * RENAME command now keeps key expiration